+++ R/package.R
@@ -13,9 +13,7 @@
 #' @section Calculations:
 #' \itemize{
 #'   \item{\code{\link{calculate_distance}}: Calculate pairwise distances between two matrices}
-#'   \item{\code{\link{calculate_arithmetic_mean}}: Calculate the arithmetic between various sorts of objects}
-#'   \item{\code{\link{calculate_geometric_mean}}: Calculate the geometric between various sorts of objects}
-#'   \item{\code{\link{calculate_harmonic_mean}}: Calculate the harmonic between various sorts of objects}
+#'   \item{\code{\link{calculate_mean}}: Calculate a (weighted) mean between vectors or a list of vectors; supports the arithmetic, geometric and harmonic mean}
 #'   \item{\code{\link{project_to_segments}}: Project a set of points to to set of segments}
 #' }
 #'