+++ inst/NEWS.md
@@ -8,7 +8,7 @@
  
  * FUNCTIONALITY: Added `update_news()` and `recent_news()`. So meta!
  
- * FUNCTIONALITY: Moved `calculate_{geometric,harmonic,arithmetic}_mean()` from dyneval to dynutils.
+ * FUNCTIONALITY: Moved `calculate_mean()` from dyneval to dynutils.
  
 # dynutils 1.0.1 (2019-02-04)